requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==2015.4.28
Flask==0.10.1
html5lib==0.99999
httplib2==0.9.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
oauth==1.0.1
oauth2client==1.4.11
pyasn1==0.1.7
pyasn1-modules==0.0.5
PyJWT==1.3.0
requests==2.7.0
rsa==3.1.4
six==1.9.0
SQLAlchemy==0.8.4
tornado==4.2
Werkzeug==0.10.4
